+++ R/soltimmar.R
@@ -244,13 +244,14 @@ sebms_sunmean_data <- function(year = 2018:2022, months = 4:9, per_month = FALSE
 
 #' Create an Image with Sun Hour Data
 #'
-#' Create a data frame from e.g. [sebms_sunhours_data()] and creates an raster
-#' image.
+#' Create a data frame from e.g. [sebms_sunhours_data()] and creates an raster image.
 #'
 #' @inheritParams sebms_sunhours_data
 #' @param df optional; a dataframe created by [sebms_sunhours_data()]
-#' @param sunvar the variable to calculate colours on, `total_sunH` or
-#'   `mean_sunH`
+#' @param filepath a path to the directory where the plots should be saved; default to
+#'   working directory
+#' @param tag an optional tag that is added to the end of the file name
+#' @param sunvar the variable to calculate colours on, `total_sunH` or `mean_sunH`
 #' @param legends logical; if you want a legend to the figures (default: FALSE)
 #'
 #' @importFrom lubridate year today
@@ -260,10 +261,10 @@ sebms_sunmean_data <- function(year = 2018:2022, months = 4:9, per_month = FALSE
 #' @importFrom glue glue
 #' @importFrom scales squish
 #'
-#' @return a figure saved as a png with the sunhours in coloour from, high (red)
-#'   to low (blue)
+#' @return a figure saved as a png with the sunhours in coloour from, high (red) to low
+#'   (blue)
 #' @export
-sebms_sunhour_plot <- function(year = lubridate::year(lubridate::today())-1, df, sunvar = total_sunH, months = 4:9, per_month = FALSE, per_day = FALSE, legends = FALSE) {
+sebms_sunhour_plot <- function(year = lubridate::year(lubridate::today())-1, df, filepath = getwd(), tag = NULL, sunvar = total_sunH, months = 4:9, per_month = FALSE, per_day = FALSE, legends = FALSE) {
   
   if (missing(df) && length(months) < 6 && per_month == FALSE) {
     warning("THIS FIGURE WILL LOOK VERY BLUE (LOW NR HOURS) AS IT IS OPTIMIZED FOR THE SUM OF SUNHOURS OVER 6 SUMMER MONTH\n  USE 'per_month = TRUE' TO GET VALUES PER MONTH\n  OR SET 'months=4:9'")
@@ -372,7 +373,17 @@ sebms_sunhour_plot <- function(year = lubridate::year(lubridate::today())-1, df,
       ungroup() %>% 
       mutate(plots = map(data, sunHplot, .progress = "Create sunhour figures"))
     
-    walk2(ggs$plots, ggs$Year, ~sebms_ggsave(.x, "Sweden", width = 6, height = 12.67, weathervar = glue("Sunhours_{.y}")))
+    #set tag
+    if (is.null(tag)) {
+      tag = ""
+    }else {
+      tag = glue("_{tag}")
+    }
+    #set filepath
+    filepath <- normalizePath(filepath)
+    
+    
+    walk2(ggs$plots, ggs$Year, ~sebms_ggsave(.x, glue("{filepath}/Sweden"), width = 6, height = 12.67, weathervar = glue("Sunhours_{.y}{tag}")))
     
     return(ggs$plots) 
   }
@@ -402,7 +413,7 @@ sebms_sunhour_plot <- function(year = lubridate::year(lubridate::today())-1, df,
 #'
 #' @return a figure that shows diffeence in sunhours
 #' @export
-sebms_sundiff_plot <- function(year = lubridate::year(lubridate::today())-1, df, months = 4:9, per_month = FALSE, legends = FALSE, per_day = FALSE) {
+sebms_sundiff_plot <- function(year = lubridate::year(lubridate::today())-1, df, filepath = getwd(), tag = NULL, months = 4:9, per_month = FALSE, legends = FALSE, per_day = FALSE) {
   
   if(missing(df)) {
     message("Please be pacient...")
@@ -496,8 +507,16 @@ sebms_sundiff_plot <- function(year = lubridate::year(lubridate::today())-1, df,
       ungroup() %>% 
       mutate(plots = map(data, sunDiffplot, .progress = "Create sunhour diff figures"))
     
+    #set tag
+    if (is.null(tag)) {
+      tag = ""
+    }else {
+      tag = glue("_{tag}")
+    }
+    #set filepath
+    filepath <- normalizePath(filepath)
     
-    walk2(ggs$plots, ggs$Year,  ~sebms_ggsave(.x, "Sweden", width = 6, height = 12.67, weathervar = glue("SunhourDiff_{.y}")), .progress = "Saving figures as png...")
+    walk2(ggs$plots, ggs$Year,  ~sebms_ggsave(.x, glue("{filepath}/Sweden"), width = 6, height = 12.67, weathervar = glue("SunhourDiff_{.y}{tag}")), .progress = "Saving figures as png...")
     
     return(ggs$plots)
   }